requirements.txt
backports.tempfile==1.0
cachetools==3.0.0
django>=1.11.0,<1.12.0
elasticsearch==5.3.0
enum==0.4.6
fs==2.0.16
ipython==5.7.0
jsonschema==2.6.0
Markdown==2.6.11
prometheus_client==0.2.0
psutil==5.4.6
pymongo==3.7.1
python-dateutil==2.7.3
python-dxf==7.4.1
pytz==2018.4
requests[security]==2.20.1
rfc3987==1.3.7
strict-rfc3339==0.7
unicodecsv==0.9.0
urllib3==1.23
webapp2==2.5.2
WebOb==1.8.2
git+https://github.com/flywheel-io/gears.git@v0.1.8#egg=gears
google-cloud-bigquery==1.5.0
